test harness works. eg acq2106_133> TX=1000 RX=0 US=10000 /mnt/local/wrs_testharness >/dev/null acq1102_009> RX=1000 TX=0 /mnt/local/wrs_testharness >/dev/null
TX=1000, RX=1000 : job done!
So we _could_ run at 100Hz ?. Well, maybe, we still need to check that the
packet is received in time to load the WRTTx register.
Which, to be fair, maybe it is, because we're not falling behind.

OTOH, it also failed, works better at 50Hz.

NEXT: goRT()
 On branch wrs_rw_tester
	modified:   acq400_drv.c
	modified:   acq400_wrdrv.c
	modified:   wrs_testharness.cpp