Modify how ACQ465 is initialized
Initialization is now called explicitly during boot.
It was previously called as part of a job run upon the completion of rc.user file.
This was leading to multiple writes to registers that the module is sensitive to.

The job is now defined during boot, explicitly called during boot, and explicitly
called again each time a sync_role is performed.