Modify get_vr to fix bug with modules that don't have a matching case
Fix for bug in choose_vap when it is presented with a FRU_PART_NUM that
does not match any of the case statements and falls through to the
case that parses the part number looking for a -xxV.

get_vr was returning two values each time (voltage and a zero) which was
causing next part `if [ $VSPEC -gt 0 ];` to fail. On a part ending `-5V`...
`VSPEC=5 0` substituting into `[ 5 0 -gt 0 ]` which generated an `operand 0`
error.

The shell on boxes was returning from a subshell within `get_vr` and then
returning zero, which returned both collected arguments.

New `get_vr` function doesn't exhibit this behaviour.